refactor(ci): group upgrade-check scripts under ci/upgrade-check
Mirrors the ci/abi-compat/ structure: check.ts (CI entrypoint),
lib.ts (core logic), list.ts (local multi-package report), hints.ts
(domain config).
