+++ ci/upgrade-check/hints.ts
